requirements.txt
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
streamlit==1.40.2
streamlit==1.40.1
streamlit==1.37.1
